update doi rights properties (#29)
This adds the following properties to the Rights field of a doi:
rightsIdentifier, rightsIdentifierScheme, schemeUri and lang. This was
done to make the rights properties up to date with that of data cites at
the time of writing.
Buttons to automatically fill out our two most used licenses were also
added to the gui.

---------

Co-authored-by: Jonathan Lindahl <jonathan.lindahl@nateko.lu.se>
Co-authored-by: Andrew Debevec <andrew.debevec@nateko.lu.se>